Changed to allocate bytes not longs.
Thanks jnicholls for this message.

"You're allocating 4x the amount of memory by specifying sizeof(size_t) instead of sizeof(uint8_t)"

+++ src/utils/alloc.c
@@ -56,7 +56,7 @@ void *nn_alloc_ (size_t size, const char *name)
 {
     uint8_t *chunk;
 
-    chunk = calloc (sizeof (struct nn_alloc_hdr) + size, sizeof(size_t));
+    chunk = calloc (sizeof (struct nn_alloc_hdr) + size, sizeof(uint8_t));
     if (!chunk)
         return NULL;
 
@@ -130,7 +130,7 @@ void nn_alloc_term (void)
 
 void *nn_alloc_ (size_t size)
 {
-    return calloc (size, sizeof(size_t));
+    return calloc (size, sizeof(uint8_t));
 }
 
 void *nn_realloc (void *ptr, size_t size)
